Revert "move some pdf integration tests to extended tests"
This reverts commit 48ca84dcfa46afd1eba69c073405924fc5da65df.
